TB-327 Update FhirResourcesLoader and JSONViewer components
